fix: (quick)saving during an event still sometimes resulted in breakage
Previously, we performed the save immediately, even when calling the
save function from deep inside the Lua interpreter. Because we don't
(can't) save the Java stack, this could lead to issues.

The fix is to perform the save/load at the very start of the frame,
where our call stack should always be equivalent.